GEOMESA-996 Rename command line commands
Renamed and documented changes to command line tools commands. Commands now follow the 'verb-noun' structure and are (mostly) standardized.
