+++ docs/tutorials/geomesa-raster.rst
@@ -62,7 +62,7 @@ Once the GeoMesa tools have been deployed you will be able to ingest the prepare
 
 .. code-block:: bash
 
-    $ geomesa ingestraster -u username -p password \
+    $ geomesa ingest-raster -u username -p password \
         -t natearth -f "/path/to/pyramid/1/" -F geotiff
 
 The ingest needs to be repeated per level of the pyramid generated by
